Revert "Initial break up of functions.php"
